FIX[#2353] Silently
git-svn-id: svn://localhost/svn/uclinux-dist/trunk@4606 3e2a5524-39e8-4cb9-84a9-c49da69cba3a

+++ linux-2.6.x/include/asm-blackfin/types.h
@@ -57,7 +57,7 @@ typedef unsigned long long u64;
 /* Dma addresses are 32-bits wide.  */
 
 typedef u32 dma_addr_t;
-typedef u32 dma64_addr_t;
+typedef u64 dma64_addr_t;
 
 #endif				/* __ASSEMBLY__ */
